requirements.txt
streamlit
pandas
scikit-learn
requests
beautifulsoup4
python-dotenv
pyserial